Safari-609.1.11.1/LayoutTests/accessibility/poorly-formed-aria-table.html

<!DOCTYPE HTML>
<html>
<body>
<script src="../resources/js-test-pre.js"></script>

<div id="content">

<div role="grid" id="grid">
  <div role="application">
     <div role="application">
         <div role="row" id="row1">
            <div role="gridcell">1</div>
         </div>
         <div role="text">a</div>
         <div role="row" id="row2">
            <div role="gridcell">1</div>
         </div>
     </div>
  </div>
</div>

</div>

<div id="console"></div>
<script>

description("This tests that if an ARIA table does not report the rows as its direct children, the table will still correctly report the rows.")

if (window.testRunner && window.accessibilityController) {
 
    var grid = accessibilityController.accessibleElementById("grid");
    var row1 = accessibilityController.accessibleElementById("row1");
    var row2 = accessibilityController.accessibleElementById("row2");
    
 
    shouldBeTrue("grid.rowAtIndex(0).isEqual(row1)");
    shouldBeTrue("grid.rowAtIndex(1).isEqual(row2)");
    shouldBe("grid.rowCount", "2");

    document.getElementById("content").style.visibility = 'hidden';    
}
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
